Broke observed params out for refactor
